revamp notification system
